refactor(center): 사용하지 않는 기존 Center 기능 제거
