+++ src/main/java/com/somemore/domains/center/repository/center/CenterRepository.java
@@ -1,13 +1,13 @@
 package com.somemore.domains.center.repository.center;
 
+import com.somemore.center.repository.record.CenterOverviewInfo;
 import com.somemore.domains.center.domain.Center;
-import com.somemore.domains.center.repository.mapper.CenterOverviewInfo;
-
 import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
 
 public interface CenterRepository {
+
     Center save(Center center);
 
     boolean existsById(UUID id);
