+++ src/test/java/com/somemore/domains/center/service/query/CenterQueryServiceTest.java
@@ -1,25 +1,20 @@
 package com.somemore.domains.center.service.query;
 
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.assertThatCode;
+import static org.assertj.core.api.Assertions.assertThatExceptionOfType;
+
 import com.somemore.domains.center.domain.Center;
-import com.somemore.domains.center.domain.PreferItem;
-import com.somemore.domains.center.dto.response.CenterProfileResponseDto;
-import com.somemore.domains.center.repository.center.CenterJpaRepository;
 import com.somemore.domains.center.repository.center.CenterRepository;
-import com.somemore.domains.center.repository.mapper.CenterOverviewInfo;
-import com.somemore.domains.center.repository.preferitem.PreferItemJpaRepository;
 import com.somemore.global.exception.BadRequestException;
 import com.somemore.global.exception.ExceptionMessage;
 import com.somemore.support.IntegrationTestSupport;
 import jakarta.transaction.Transactional;
+import java.util.UUID;
 import org.assertj.core.api.ThrowableAssert;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
-
-import java.util.List;
-import java.util.UUID;
-
-import static org.assertj.core.api.Assertions.*;
 
 @Transactional
 class CenterQueryServiceTest extends IntegrationTestSupport {
@@ -29,76 +24,6 @@ class CenterQueryServiceTest extends IntegrationTestSupport {
 
     @Autowired
     private CenterRepository centerRepository;
-
-    @Autowired
-    private CenterJpaRepository centerJpaRepository;
-    
-    @Autowired
-    private PreferItemJpaRepository preferItemRepository;
-
-    @DisplayName("기관 Id로 기관 프로필을 조회할 수 있다. (service)")
-    @Test
-    void getCenterProfileById() {
-        // given
-        Center center = createCenter();
-        Center foundCenter = centerRepository.save(center);
-
-        PreferItem preferItem = PreferItem.create(foundCenter.getId(), "어린이 동화책");
-        PreferItem preferItem1 = PreferItem.create(foundCenter.getId(), "간식");
-        preferItemRepository.saveAll(List.of(preferItem, preferItem1));
-
-        // when
-        CenterProfileResponseDto responseDto = centerQueryService.getCenterProfileByCenterId(foundCenter.getId());
-
-        // then
-        assertThat(responseDto)
-                .extracting(
-                        "centerId",
-                        "name",
-                        "contactNumber",
-                        "imgUrl",
-                        "introduce",
-                        "homepageLink"
-                )
-                .containsExactly(
-                        foundCenter.getId(),
-                        "기본 기관 이름",
-                        "010-1234-5678",
-                        "http://example.com/image.jpg",
-                        "기관 소개 내용",
-                        "http://example.com"
-                );
-
-        assertThat(responseDto.preferItems())
-                .hasSize(2)
-                .extracting("itemName")
-                .containsExactly("어린이 동화책", "간식");
-    }
-
-    @DisplayName("기관 Id들로 기관의 오버뷰 정보를 조회할 수 있다.")
-    @Test
-    void getCenterOverviewsByIds() {
-        //given
-        Center center = createCenter();
-        Center center1 = createCenter();
-        Center center2 = createCenter();
-        centerJpaRepository.saveAll(List.of(center, center1, center2));
-        List<UUID> ids = List.of(center.getId(),center1.getId(), center2.getId());
-
-        //when
-        List<CenterOverviewInfo> responseDtos = centerQueryService.getCenterOverviewsByIds(ids);
-
-        //then
-        assertThat(responseDtos)
-                .isNotNull()
-                .hasSize(3)
-                .extracting("centerId", "centerName", "imgUrl")
-                .containsExactlyInAnyOrder(
-                        tuple(center.getId(), center.getName(), center.getImgUrl()),
-                        tuple(center1.getId(), center1.getName(), center1.getImgUrl()),
-                        tuple(center2.getId(), center2.getName(), center2.getImgUrl())
-                );
-    }
 
     @DisplayName("존재하지 않는 기관 ID를 검증할 수 있다.")
     @Test
@@ -130,7 +55,8 @@ class CenterQueryServiceTest extends IntegrationTestSupport {
         Center savedCenter = centerRepository.save(center);
 
         // when
-        ThrowableAssert.ThrowingCallable callable = () -> centerQueryService.validateCenterExists(savedCenter.getId());
+        ThrowableAssert.ThrowingCallable callable = () -> centerQueryService.validateCenterExists(
+                savedCenter.getId());
 
         // then
         assertThatCode(callable).doesNotThrowAnyException();
@@ -155,7 +81,8 @@ class CenterQueryServiceTest extends IntegrationTestSupport {
     void getNameByNonExistentId() {
         // given
         // when
-        ThrowableAssert.ThrowingCallable callable = () -> centerQueryService.getNameById(UUID.randomUUID());
+        ThrowableAssert.ThrowingCallable callable = () -> centerQueryService.getNameById(
+                UUID.randomUUID());
 
         // then
         assertThatExceptionOfType(BadRequestException.class)
